+++ comments.php
@@ -23,8 +23,41 @@ if ( comments_open() || pings_open() ) {
 
 	if ( $comments ) {
 		//echo '<hr class="styled-separator is-style-wide" aria-hidden="true" />';
+
+		$i = 0;
+		$total = 0;
+		$ratingCount = 0;
+		$aggregateRating;
+		foreach( $comments as $comment ){
+			$rate = get_comment_meta( $comment->comment_ID, 'rating', true );
+			if( isset( $rate ) && '' !== $rate ) {
+				$i++;
+				$total += $rate;
+			}
+		}
+
+		if ( 0 === $i ) {
+			//return false;
+		} else {
+			$aggregateRating =  round( $total / $i, 1 );?>
+
+
+<?php
+		}
 	}
 ?>
+
+	<span itemprop="interactionStatistic" itemscope itemtype="https://schema.org/InteractionCounter">
+    <meta itemprop="interactionType" content="https://schema.org/CommentAction"/>
+    <meta itemprop="userInteractionCount" content="<?php echo count($comments);?>" />
+	</span>
+
+		<span itemprop="aggregateRating" itemscope itemtype="https://schema.org/AggregateRating">
+  		<meta itemprop="ratingValue" content="<?php echo  $aggregateRating; ?>" />
+  		<meta itemprop="reviewCount" content="<?php echo  $i; ?>" />
+      <meta itemprop="worstRating" content = "0">
+      <meta itemprop="BestRating" content = "5">
+		</span>
   <div  class="lyt">
     <div class="cxb">
       <?php